# belt-service/src/main/java/com/coderdojo/zen/belt/model/Belt.java
package com.coderdojo.zen.belt.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

@Document(value = "belt")
@AllArgsConstructor
@NoArgsConstructor
@Builder
@Data
public class Belt {

    @Id
    private String id;

    private String name;

    private String description;

    private String imageURL;

}