# belt-service/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Belt Service API
  description: This is a sample Foobar server created using springdocs - a library
    for OpenAPI 3 with spring boot.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8080
    description: Generated server url
...
paths:
  /api/v1/belts:
    get:
      tags:
        - Belts
      operationId: all
      summary: Get all belts
      parameters:
          - name: category
            in: query
            description: ID of the object to fetch
            required: false
            schema:
              type: string
              enum:
                - programming
                - soft_skills
                - events
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelBelt'
    post:
      tags:
        - Belts
      operationId: newBelt
      summary: Create belt
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Belt'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EntityModelBelt'
...
components:
  schemas:
    Belt:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        status:
          type: string
          enum:
            - IN_PROGRESS
            - COMPLETED
            - CANCELLED
    EntityModelBelt:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        status:
          type: string
          enum:
            - IN_PROGRESS
            - COMPLETED
            - CANCELLED
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean
    CollectionModelEntityModelBelt:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            beltList:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelBelt'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'